find guests done
